requirements.txt
# Core API framework
fastapi==0.115.12
uvicorn==0.34.0

# Data validation
pydantic==2.11.2
pydantic-settings==2.8.1

# Database (SQLite for local dev, PostgreSQL driver for production)
sqlalchemy==2.0.40
aiosqlite==0.20.0
asyncpg==0.30.0

# File I/O
python-multipart==0.0.20
python-dotenv==1.1.0
aiofiles==24.1.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# ML & AI
sentence-transformers==4.1.0
scikit-learn==1.6.1
numpy==2.2.4
faiss-cpu==1.10.0

# HTTP client
requests==2.32.3
huggingface_hub==0.30.2

# LangChain (pinned to stable local version)
langchain==0.3.23
langchain-community==0.3.21
langchain-huggingface==0.1.2

# Optional: Async tasks (comment out if no Redis available locally)
# celery==5.3.6
# redis==5.0.1